Fix the "double-drop" issue
This is caused by failing to reset the drop timer (`ticksUntilDrop`)
when the user moves the piece down. What happens is the user moves the
piece down just before the timer hits zero, and then "gravity" moves it
down immediately after. This creates a jarring experience for the
player.

The solution is to reset the timer regardless of whether it was the user
or gravity that moved the piece down.